+++ scraper/src/parseResponse.js
