+++ scraper/src/updateLucky.js
